move components directory and rename pages components
