+++ pages/Contact.tsx
@@ -1,6 +1,6 @@
 import Link from 'next/link'
 import React from 'react'
-import { Hero } from './components/organisms/Hero'
+import { Hero } from '../components/organisms/Hero'
 
 export default function Contact() {
   return (
